Implementando exceções para os demais métodos

+++ src/main/java/com/rafa/dscommerce/services/ProductService.java
@@ -3,11 +3,15 @@ package com.rafa.dscommerce.services;
 import com.rafa.dscommerce.dto.ProductDTO;
 import com.rafa.dscommerce.entities.Product;
 import com.rafa.dscommerce.repositories.ProductRepository;
+import com.rafa.dscommerce.services.exceptions.DatabaseException;
 import com.rafa.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -38,16 +42,27 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto){
-        // Objeto monitorado pela JPA
-        Product product = repository.getReferenceById(id);
-        copyDtoToEntity(dto, product);
-        product = repository.save(product);
-        return new ProductDTO(product);
+        try{
+            // Objeto monitorado pela JPA
+            Product product = repository.getReferenceById(id);
+            copyDtoToEntity(dto, product);
+            product = repository.save(product);
+            return new ProductDTO(product);
+        }catch(EntityNotFoundException e){
+            throw new ResourceNotFoundException("Produto não encontrado");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
-        repository.deleteById(id);
+        if(!repository.existsById(id)){
+            throw  new ResourceNotFoundException("Produto não encontrado");
+        }
+        try{
+            repository.deleteById(id);
+        }catch(DataIntegrityViolationException e){
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product product) {
